% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiRhino.R
\name{hiRhino}
\alias{hiRhino}
\title{Say hi to the animals}
\usage{
hiRhino(n = 1)
}
\arguments{
\item{n}{Number of times to say hello}
}
\value{
A nice greeting
}
\description{
Say hi to the animals
}
\examples{
hiRhino()
hiRhino(2)
}
