% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiRhino.R
\name{hiRhino}
\alias{hiRhino}
\title{Say hi to the Rhinos.}
\usage{
hiRhino(n = 1)
}
\arguments{
\item{n}{A number.}
}
\value{
A number \code{n} greetings to the animals.
}
\description{
Say hi to the Rhinos.
}
